+++ glean-core/rlb/examples/long-running.rs
@@ -33,12 +33,7 @@ pub mod glean_metrics {
 struct FakeUploader;
 
 impl net::PingUploader for FakeUploader {
-    fn upload(
-        &self,
-        _url: String,
-        _body: Vec<u8>,
-        _headers: Vec<(String, String)>,
-    ) -> net::UploadResult {
+    fn upload(&self, _upload_request: net::PingUploadRequest) -> net::UploadResult {
         thread::sleep(time::Duration::from_millis(100));
         net::UploadResult::http_status(200)
     }
@@ -46,7 +41,7 @@ impl net::PingUploader for FakeUploader {
 
 #[allow(non_upper_case_globals)]
 pub static PrototypePing: Lazy<PingType> =
-    Lazy::new(|| PingType::new("prototype", true, true, true, vec![]));
+    Lazy::new(|| PingType::new("prototype", true, true, true, true, vec![]));
 
 fn main() {
     env_logger::init();