+++ CHANGELOG.md
@@ -1,13 +1,25 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v57.0.0...main)
+[Full changelog](https://github.com/mozilla/glean/compare/v58.0.0...main)
+
+# v58.0.0 (2024-02-29)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v57.0.0...v58.0.0)
+
+* General
+  * Update `glean_parser` to v13.0.0 ([release notes](https://github.com/mozilla/glean_parser/releases/tag/v13.0.0))
+* Rust
+  * New metric type: Object ([#2489](https://github.com/mozilla/glean/pull/2489))
+  * BREAKING CHANGE: Support pings without `{client|ping}_info` sections ([#2756](https://github.com/mozilla/glean/pull/2756))
+* Android
+  * Upgrade Android NDK to r26c ([#2745](https://github.com/mozilla/glean/pull/2745))
 
 # v57.0.0 (2024-02-12)
 
 [Full changelog](https://github.com/mozilla/glean/compare/v56.1.0...v57.0.0)
 
 * General
-  * Added an experimental event listener API ([#2719](https://github.com/mozilla/glean/pull/2719)) 
+  * Added an experimental event listener API ([#2719](https://github.com/mozilla/glean/pull/2719))
 * Android
   * BREAKING CHANGE: Update JNA to version 5.14.0. Projects using older JNA releases may encounter errors until they update. ([#2727](https://github.com/mozilla/glean/pull/2727))
   * Set the target Android SDK to version 34 ([#2709](https://github.com/mozilla/glean/pull/2709))